+++ generators/messaging-adapter-core/src/main/java/uk/gov/justice/services/core/json/JsonValidationLoggerHelper.java
@@ -2,11 +2,9 @@ package uk.gov.justice.services.core.json;
 
 import javax.json.JsonObject;
 
-import org.everit.json.schema.ValidationException;
-
 public interface JsonValidationLoggerHelper {
 
-    String toValidationTrace(final ValidationException validationException);
+    String toValidationTrace(final JsonSchemaValidatonException validationException);
 
-    JsonObject toJsonObject(final ValidationException validationException);
+    JsonObject toJsonObject(final JsonSchemaValidatonException validationException);
 }